backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyGithub==2.1.1
notion-client==2.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
python-dateutil==2.8.2
